Added some modifications
- Using new DriftPHP kernel feature that allows to create loop interval
service calls as configuration
- Made a controller abstraction
- Changed DBAL named injection to allow multiple DBAL connection
injections
