+++ Plugin/DBAL/Resources/config/domain.yml
@@ -11,6 +11,7 @@ services:
             string $searchesTable: "%apisearch_plugin.dbal.searches_table%"
             bool $locatorEnabled: "%apisearch_plugin.dbal.locator_enabled%"
             int $loopPushInterval: "%apisearch_plugin.dbal.loop_push_interval%"
+
             $persistentUsageRepository: '@Apisearch\Plugin\DBAL\Domain\UsageRepository\DBALUsageRepository'
 
     Apisearch\Plugin\DBAL\Domain\:
@@ -45,6 +46,9 @@ services:
     #
     Apisearch\Plugin\DBAL\Domain\UsageRepository\DBALUsageRepository:
     Apisearch\Plugin\DBAL\Domain\UsageRepository\ChunkUsageRepository:
+        tags:
+            - {name: "periodic_timer", interval: "%apisearch_plugin.dbal.loop_push_interval%", method: "flushLines"}
+
     Apisearch\Server\Domain\Repository\UsageRepository\UsageRepository:
         alias: Apisearch\Plugin\DBAL\Domain\UsageRepository\ChunkUsageRepository
