+++ composer.json
@@ -10,7 +10,7 @@
     }
   ],
   "require": {
-    "php": ">=7.3",
+    "php": ">=7.4",
     "ext-json": "*",
 
     "symfony/framework-bundle": "^5.0",
@@ -21,7 +21,7 @@
     "symfony/event-dispatcher" : "^5.0",
     "symfony/dotenv" : "^5.0",
 
-    "drift/http-kernel": "0.1.*, >=0.1.11",
+    "drift/http-kernel": "dev-master as 0.1.12",
     "drift/server": "dev-master as 0.1.14",
     "drift/react-functions": "0.1.*, >=0.1.1",
     "drift/react-streams": "0.1.*",
@@ -41,7 +41,7 @@
     "mmoreram/base-bundle": "^2.0",
     "mmoreram/symfony-bundle-dependencies": "^2.0",
 
-    "apisearch-io/php-client": "0.1.*, >=0.1.30",
+    "apisearch-io/php-client": "0.1.*, >=0.1.31",
     "apisearch-io/apisearch-bundle": "0.1.*, >=0.1.16",
 
     "ruflin/elastica": "^7.0",
